refactor: change chatCard Design
