+++ packages/gdsc-dju-admin/src/components/molecules/ChatCard/styled.ts
@@ -1,13 +1,13 @@
 import styled, { css } from 'styled-components';
 
 export const ChatCardInner = styled.div<{ isUser: boolean }>`
-  padding: 10px;
+  padding: 10px 16px;
   width: fit-content;
   min-width: 100px;
   max-width: 200px;
   box-shadow: 0 2px 3px ${({ theme }) => theme.colors.grey300};
   margin-top: 10px;
-  border-radius: 10px;
+  border-radius: 4px 16px 16px 16px;
   display: flex;
   flex-direction: column;
   flex-wrap: wrap;
@@ -18,6 +18,7 @@ export const ChatCardInner = styled.div<{ isUser: boolean }>`
     css`
       border: 1px solid ${({ theme }) => theme.colors.blue600};
       box-shadow: ${({ theme }) => theme.colors.boxShadow100};
+      border-radius: 16px 16px 4px 16px;
     `}
 `;
 export const ChatUser = styled.p`
