+++ packages/gdsc-dju-admin/src/components/organisms/ApplicantChatContainer/styled.ts
