+++ packages/gdsc-dju-admin/src/components/organisms/modal/ApplicantModal/index.tsx
@@ -5,7 +5,7 @@ import { IApplicantTypeWithID } from '../../../../types/applicant';
 import { getApplicant } from '../../../../utils/applicantsHandler';
 import { modalVariants } from '../../../animations/modalVariants';
 import { ClearButton } from '../../../atoms/ModalButton';
-import ApplicantChatSection from '../../../molecules/ApplicantChatContainer';
+import ApplicantChatSection from '../../../organisms/ApplicantChatContainer';
 import ApplicantInfoCard from '../../../molecules/ApplicantInfoCard';
 import ApplicantInfoState from '../../../molecules/ApplicantInfoState';
 import Application from '../../../molecules/Applicantion';
